factor out withclient functionality so lantern-android can use it
