+++ src/github.com/getlantern/flashlight/flashlight.go
@@ -5,12 +5,10 @@ import (
 	"flag"
 	"fmt"
 	"math/rand"
-	"net/http"
 	"os"
 	"reflect"
 	"runtime"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/getlantern/fronted"
@@ -30,7 +28,7 @@ import (
 	"github.com/getlantern/flashlight/statreporter"
 	"github.com/getlantern/flashlight/statserver"
 	"github.com/getlantern/flashlight/ui"
-	"github.com/getlantern/flashlight/util"
+	"github.com/getlantern/flashlight/withclient"
 )
 
 var (
@@ -93,81 +91,6 @@ func _main() {
 	os.Exit(0)
 }
 
-// The following are mostly meant to
-//   (i)   keep track of directly fronted Dialers, closing them when they
-//         get superseded by a config update.
-//   (ii)  avoid creating more Dialers than needed (one per config update).
-//   (iii) make these dialers available to the config custom poll, without
-//         introducing globals.
-//   (iv)  present a uniform interface in the server side, currently for
-//         the benefit of the config custom poll.
-//
-// Thus, most of the functionality is meant for the client side, and we
-// use mostly placeholders in the server side.  Because of this and for
-// agility of exposition, the comments below refer to the client side.
-
-type clientWithCloseThunk struct {
-	// A particular reference to the directly fronted http.Client (aka fronter).
-	client *http.Client
-	// Placeholder for any cleanup required when this reference is no longer
-	// needed.  We keep track of currently used references with a WaitGroup,
-	// and we use this function to signal we're Done() with this one.
-	close func()
-}
-
-type clientMaker struct {
-	// Generates structures like the one above.  A closure over a fronted.Dialer.
-	make func() clientWithCloseThunk
-	// Placeholder for any cleanup required when this clientMaker is no longer
-	// current because we have got a new one (by a config update).
-	close func()
-}
-
-// To synchronize access to the current clientMaker.  Used as a promise that
-// can be updated.
-type makerChan chan clientMaker
-
-func newMakerChan() makerChan {
-	return make(chan clientMaker, 1)
-}
-
-// Returns the old one, if any.
-func (ch makerChan) updateMaker(c clientMaker) (ret clientMaker) {
-	select {
-	case ret = <-ch:
-	default:
-	}
-	ch <- c
-	return ret
-}
-
-func (ch makerChan) getMaker() clientMaker {
-	ret := <-ch
-	ch <- ret
-	return ret
-}
-
-// Creates a "context" function that takes care of all the bookkeeping involved
-// in making sure Dialers are kept only for as long as needed and closed as soon as
-// nobody is using them-- but no sooner.
-//
-// The function returned by this call is meant to be used like this:
-//
-// withClient(func(c *http.Client) {
-//    ... use `c` here ...
-// })
-//
-// No explicit cleanup is required in the body where `c` is used, but `c` should
-// never be used after the body has returned.  So don't assign `c` to variables
-// or data structures outside the body, don't use it inside a goroutine, etc.
-func (ch makerChan) makeWithClient() func(func(*http.Client)) {
-	return func(f func(*http.Client)) {
-		cc := ch.getMaker().make()
-		defer cc.close()
-		f(cc.client)
-	}
-}
-
 func doMain() error {
 	err := logging.Init()
 	if err != nil {
@@ -190,9 +113,9 @@ func doMain() error {
 
 	parseFlags()
 
-	mch := newMakerChan()
+	mch := withclient.NewMakerChan()
 
-	cfg, err := config.Init(mch.makeWithClient())
+	cfg, err := config.Init(mch.MakeWithClient())
 	if err != nil {
 		return fmt.Errorf("Unable to initialize configuration: %v", err)
 	}
@@ -260,7 +183,7 @@ func parseFlags() {
 }
 
 // runClientProxy runs the client-side (get mode) proxy.
-func runClientProxy(cfg *config.Config, mch makerChan) {
+func runClientProxy(cfg *config.Config, mch withclient.MakerChan) {
 	var err error
 
 	// Set Lantern as system proxy by creating and using a PAC file.
@@ -296,8 +219,8 @@ func runClientProxy(cfg *config.Config, mch makerChan) {
 	settings.Configure(cfg, version, buildDate)
 	proxiedsites.Configure(cfg.ProxiedSites)
 
-	updateClientDirectFronter(cfg, mch)
-	wdc := mch.makeWithClient()
+	withclient.UpdateClientDirectFronter(cfg, mch)
+	wdc := mch.MakeWithClient()
 	geolookup.Configure(wdc)
 	statserver.Configure(wdc)
 	analytics.Configure(cfg, false, wdc)
@@ -320,7 +243,7 @@ func runClientProxy(cfg *config.Config, mch makerChan) {
 			// XXX: wrt fronted servers, we only really care if the one with highest QOS
 			// is still the same one.
 			if oldCfg == nil || !(reflect.DeepEqual(oldCfg.Client.FrontedServers, cfg.Client.FrontedServers) && reflect.DeepEqual(oldCfg.Client.MasqueradeSets, cfg.Client.MasqueradeSets)) {
-				updateClientDirectFronter(cfg, mch)
+				withclient.UpdateClientDirectFronter(cfg, mch)
 			}
 			settings.Configure(cfg, version, buildDate)
 			logging.Configure(cfg, version, buildDate)
@@ -337,40 +260,11 @@ func runClientProxy(cfg *config.Config, mch makerChan) {
 	}()
 }
 
-func updateClientDirectFronter(cfg *config.Config, mch makerChan) {
-	log.Debug("Updating client direct fronter")
-	hqfd := cfg.Client.HighestQOSFrontedDialer()
-	if hqfd == nil {
-		log.Errorf("No fronted dialer available, not enabling geolocation, stats or analytics")
-		return
-	}
-	// An *http.Client that uses the highest QOS dialer.
-	hqfdClient := hqfd.NewDirectDomainFronter()
-	wg := sync.WaitGroup{}
-	old := mch.updateMaker(
-		clientMaker{
-			make: func() clientWithCloseThunk {
-				wg.Add(1)
-				return clientWithCloseThunk{
-					client: hqfdClient,
-					close:  wg.Done,
-				}
-			},
-			close: func() {
-				wg.Wait()
-				hqfd.Close()
-			}})
-	if old.close != nil {
-		log.Debug("Closing old dialer")
-		go old.close()
-	}
-}
-
 // Runs the server-side proxy
-func runServerProxy(cfg *config.Config, mch makerChan) {
+func runServerProxy(cfg *config.Config, mch withclient.MakerChan) {
 	useAllCores()
 
-	updateServerConfigClient(cfg, mch)
+	withclient.UpdateServerConfigClient(cfg, mch)
 	pkFile, err := config.InConfigDir("proxypk.pem")
 	if err != nil {
 		log.Fatal(err)
@@ -403,7 +297,7 @@ func runServerProxy(cfg *config.Config, mch makerChan) {
 		for {
 			cfg := <-configUpdates
 			if cfg.CloudConfigCA != oldca {
-				updateServerConfigClient(cfg, mch)
+				withclient.UpdateServerConfigClient(cfg, mch)
 				oldca = cfg.CloudConfigCA
 			}
 			statreporter.Configure(cfg.Stats)
@@ -422,24 +316,6 @@ func runServerProxy(cfg *config.Config, mch makerChan) {
 	if err != nil {
 		log.Fatalf("Unable to run server proxy: %s", err)
 	}
-}
-
-func updateServerConfigClient(cfg *config.Config, mch makerChan) {
-	client, err := util.HTTPClient(cfg.CloudConfigCA, "")
-	if err != nil {
-		log.Errorf("Couldn't create http.Client for fetching the config")
-		return
-	}
-	doNothing := func() {}
-	ret := clientWithCloseThunk{
-		client: client,
-		close:  doNothing,
-	}
-	mch.updateMaker(
-		clientMaker{
-			make:  func() clientWithCloseThunk { return ret },
-			close: doNothing,
-		})
 }
 
 func useAllCores() {
